Refactor: 모달 스크롤 로직 분리
