+++ inflow/src/components/modals/CrudModal.vue
@@ -26,6 +26,7 @@
 </template>
 
 <script setup>
+import { onMounted, onBeforeUnmount } from 'vue';
 import XmarkButton from '../buttons/XmarkButton.vue';
 import ModalBackground from '../modals/ModalBackground.vue';
 import ModalItem from '../modals/ModalItem.vue';
@@ -47,6 +48,40 @@ const emit = defineEmits(['close']);
 const closeModal = () => {
   emit('close');
 };
+
+const preventScroll = (e) => {
+  const modal = document.querySelector(".modal-background");
+
+  // 모달이 존재하고, 이벤트의 타겟이 모달 내부에 있으면 스크롤 허용
+  if (modal && modal.contains(e.target)) {
+    return; // 모달 내부의 스크롤은 차단하지 않음
+  }
+
+  // 모달 외부에서는 스크롤 차단
+  e.preventDefault();
+};
+
+const handlePreventScrollListener = (isModalOpen) => {
+  if (isModalOpen) {
+    // 스크롤 차단
+    window.addEventListener("wheel", preventScroll, { passive: false }); // 마우스 휠
+    window.addEventListener("touchmove", preventScroll, { passive: false }); // 모바일 터치
+    document.body.style.overflow = "hidden"; // 추가적으로 body 스크롤 차단
+  } else {
+    // 스크롤 활성화
+    window.removeEventListener("wheel", preventScroll);
+    window.removeEventListener("touchmove", preventScroll);
+    document.body.style.overflow = ""; // body 스크롤 다시 활성화
+  }
+};
+
+onMounted(() => {
+  handlePreventScrollListener(true);
+})
+
+onBeforeUnmount(() => {
+  handlePreventScrollListener(false);
+})
 </script>
 
 <style scoped>
